update tests for now

+++ Tests/Support/Snapshots/edgecases-int32-int64/Sources/Entities/FormatTest.swift
@@ -6,8 +6,8 @@ import NaiveDate
 
 public struct FormatTest: Codable {
     public var integer: Int?
-    public var int32: Int32?
-    public var int64: Int64?
+    public var int32: Int?
+    public var int64: Int?
     public var number: Double
     public var float: Double?
     public var double: Double?
@@ -19,7 +19,7 @@ public struct FormatTest: Codable {
     public var uuid: UUID?
     public var password: String
 
-    public init(integer: Int? = nil, int32: Int32? = nil, int64: Int64? = nil, number: Double, float: Double? = nil, double: Double? = nil, string: String? = nil, byte: Data, binary: String? = nil, date: NaiveDate, dateTime: Date? = nil, uuid: UUID? = nil, password: String) {
+    public init(integer: Int? = nil, int32: Int? = nil, int64: Int? = nil, number: Double, float: Double? = nil, double: Double? = nil, string: String? = nil, byte: Data, binary: String? = nil, date: NaiveDate, dateTime: Date? = nil, uuid: UUID? = nil, password: String) {
         self.integer = integer
         self.int32 = int32
         self.int64 = int64
@@ -38,8 +38,8 @@ public struct FormatTest: Codable {
     public init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
         self.integer = try values.decodeIfPresent(Int.self, forKey: "integer")
-        self.int32 = try values.decodeIfPresent(Int32.self, forKey: "int32")
-        self.int64 = try values.decodeIfPresent(Int64.self, forKey: "int64")
+        self.int32 = try values.decodeIfPresent(Int.self, forKey: "int32")
+        self.int64 = try values.decodeIfPresent(Int.self, forKey: "int64")
         self.number = try values.decode(Double.self, forKey: "number")
         self.float = try values.decodeIfPresent(Double.self, forKey: "float")
         self.double = try values.decodeIfPresent(Double.self, forKey: "double")
